Refactored things a bit
